evals: run localmind test on chat, big, gemma-31b + ThinkingCap candidate
Adds four more model runs to the localmind-landing-page test and refactors
run.html generation into eval-summary.py so a summary rebuild regenerates every
run's page from disk (meta.json + check.json) — no model re-query needed.

Results (objective / decode t/s / MTP accept, cold-prefill metrics):
- coding           27/27  44.4 t/s  92%   (Qwen3.6-27B Q6_K)
- big              27/27  42.8 t/s  92%   (27B UD-Q6_K_XL, dual-V100)
- gemma-31b        27/27  28.2 t/s  32%   (Gemma-4-31B QAT Q4_K_XL)
- thinkingcap-27b  25/27  45.7 t/s  93%   (ThinkingCap fine-tune, standalone)
- chat             23/27 134.5 t/s  95%   (Qwen3.6-35B-A3B MoE — fastest decode)

The Qwen3.6-27B family (coding/big) and the ThinkingCap fine-tune produce the
most constraint-complete pages; chat is by far the fastest but misses a couple
of objective constraints. ThinkingCap matches coding's speed at parity accuracy.

ThinkingCap isn't in llama-swap, so it ran on a standalone MTP llama-server
(V100 idx2, mirroring the coding slot); its exact launch command is recorded in
meta.json via --cmd. Daily mode restored afterward.
