+++ evals/localmind-landing-page/outputs/big/meta.json
@@ -1,0 +1,63 @@
+{
+  "test": "localmind-landing-page",
+  "model_slot": "big",
+  "label": "big",
+  "model_name": "Qwen3.6-27B-UD-Q6_K_XL.gguf",
+  "model_path": "/srv/ai/models/qwen3.6-27b-mtp/Qwen3.6-27B-UD-Q6_K_XL.gguf",
+  "endpoint": "http://127.0.0.1:9090/v1/chat/completions",
+  "proxy": "http://localhost:10001",
+  "sampler": {
+    "temperature": 0.7,
+    "top_p": 0.95,
+    "top_k": 20,
+    "max_tokens": 32000
+  },
+  "load_command": "/srv/ai/src/llama.cpp/build/bin/llama-server\n--model /srv/ai/models/qwen3.6-27b-mtp/Qwen3.6-27B-UD-Q6_K_XL.gguf\n--ctx-size 262144\n--cache-type-k f16\n--cache-type-v f16\n--split-mode layer\n--parallel 1\n--spec-type draft-mtp\n--spec-draft-n-max 2\n--host 127.0.0.1 --port 10001 --n-gpu-layers 999 --flash-attn on --cont-batching --jinja --metrics",
+  "usage": {
+    "completion_tokens": 11312,
+    "prompt_tokens": 838,
+    "total_tokens": 12150,
+    "prompt_tokens_details": {
+      "cached_tokens": 0
+    }
+  },
+  "finish_reason": "stop",
+  "performance": {
+    "wall_secs": 298.7,
+    "ttft_ms": 1475.418,
+    "prefill_tps": 567.9746349847976,
+    "decode_tps": 42.79383873739593,
+    "prompt_tokens": 838,
+    "predicted_tokens": 11312,
+    "prefill_ms": 1475.418,
+    "decode_ms": 264337.118
+  },
+  "mtp": {
+    "draft_n": 7946,
+    "draft_n_accepted": 7340,
+    "accept_rate": 0.924,
+    "enabled": true
+  },
+  "server_timings": {
+    "cache_n": 0,
+    "prompt_n": 838,
+    "prompt_ms": 1475.418,
+    "prompt_per_token_ms": 1.7606420047732696,
+    "prompt_per_second": 567.9746349847976,
+    "predicted_n": 11312,
+    "predicted_ms": 264337.118,
+    "predicted_per_token_ms": 23.367849893917963,
+    "predicted_per_second": 42.79383873739593,
+    "draft_n": 7946,
+    "draft_n_accepted": 7340
+  },
+  "extraction": {
+    "extracted": true,
+    "kind": "fenced"
+  },
+  "output_file": "index.html",
+  "output_bytes": 32613,
+  "reasoning_bytes": 621,
+  "wall_secs": 298.7,
+  "created_utc": "2026-07-27T04:56:58+00:00"
+}
