+++ evals/localmind-landing-page/outputs/big/run.html
@@ -1,0 +1,128 @@
+<!doctype html>
+<html lang="en"><head><meta charset="utf-8">
+<meta name="viewport" content="width=device-width,initial-scale=1">
+<title>big — localmind-landing-page</title>
+<style>
+:root{--bg:#09090b;--surface:#13111c;--border:rgba(255,255,255,.08);
+--accent:#7c5cfc;--teal:#00d4aa;--fg:#fafafa;--muted:#71717a}
+*{box-sizing:border-box}body{margin:0;background:var(--bg);color:var(--fg);
+font:15px/1.6 -apple-system,BlinkMacSystemFont,'Segoe UI',Roboto,sans-serif;padding:32px;max-width:1100px;margin:auto}
+h1{font-size:24px;margin:0 0 2px}h2{font-size:15px;text-transform:uppercase;
+letter-spacing:.05em;color:var(--muted);margin:28px 0 10px}
+.sub{color:var(--muted);margin:0 0 20px}a{color:var(--accent);text-decoration:none}
+a:hover{text-decoration:underline}
+.grid{display:grid;grid-template-columns:repeat(auto-fit,minmax(150px,1fr));gap:12px}
+.card{background:var(--surface);border:1px solid var(--border);border-radius:10px;padding:14px 16px}
+.card .k{color:var(--muted);font-size:12px;text-transform:uppercase;letter-spacing:.04em}
+.card .v{font-size:22px;font-weight:600;margin-top:4px}
+.card .v.teal{color:var(--teal)}
+pre{background:var(--surface);border:1px solid var(--border);border-radius:10px;
+padding:14px 16px;overflow-x:auto;font-size:13px;color:#d4d4d8;white-space:pre-wrap;word-break:break-word}
+.frame{width:100%;height:70vh;border:1px solid var(--border);border-radius:10px;background:#fff}
+.pill{display:inline-block;padding:1px 8px;border-radius:999px;font-size:12px;border:1px solid var(--border)}
+.mtp-on{color:var(--teal);border-color:rgba(0,212,170,.4)}.mtp-off{color:var(--muted)}
+</style></head>
+<body>
+<h1>big</h1>
+<p class="sub">Qwen3.6-27B-UD-Q6_K_XL.gguf &middot; localmind-landing-page
+&middot; 2026-07-27T04:56:58 UTC
+&middot; <a href="../../summary.html">← all runs</a>
+&middot; <a href="index.html">output</a>
+&middot; <a href="raw.txt">raw reply</a></p>
+<div class="grid">
+<div class="card"><div class="k">Objective</div><div class="v teal">27/27</div></div>
+<div class="card"><div class="k">TTFT</div><div class="v">1475 ms</div></div>
+<div class="card"><div class="k">Prefill</div><div class="v">568.0 t/s</div></div>
+<div class="card"><div class="k">Decode</div><div class="v teal">42.8 t/s</div></div>
+<div class="card"><div class="k">Wall time</div><div class="v">298.7 s</div></div>
+<div class="card"><div class="k">Completion tok</div><div class="v">11312</div></div>
+<div class="card"><div class="k">Output size</div><div class="v">32613 B</div></div>
+<div class="card"><div class="k">MTP</div><div class="v"><span class="pill mtp-on">on</span> 92% accept</div></div>
+<div class="card"><div class="k">Finish</div><div class="v">stop</div></div>
+</div>
+<h2>Rendered output</h2>
+<iframe class="frame" src="index.html" title="rendered output"></iframe>
+
+<h2>llama.cpp load command</h2>
+<pre>/srv/ai/src/llama.cpp/build/bin/llama-server
+--model /srv/ai/models/qwen3.6-27b-mtp/Qwen3.6-27B-UD-Q6_K_XL.gguf
+--ctx-size 262144
+--cache-type-k f16
+--cache-type-v f16
+--split-mode layer
+--parallel 1
+--spec-type draft-mtp
+--spec-draft-n-max 2
+--host 127.0.0.1 --port 10001 --n-gpu-layers 999 --flash-attn on --cont-batching --jinja --metrics</pre>
+<h2>Sampler / request</h2>
+<pre>{
+  &quot;temperature&quot;: 0.7,
+  &quot;top_p&quot;: 0.95,
+  &quot;top_k&quot;: 20,
+  &quot;max_tokens&quot;: 32000
+}</pre>
+<h2>meta.json</h2>
+<pre>{
+  &quot;test&quot;: &quot;localmind-landing-page&quot;,
+  &quot;model_slot&quot;: &quot;big&quot;,
+  &quot;label&quot;: &quot;big&quot;,
+  &quot;model_name&quot;: &quot;Qwen3.6-27B-UD-Q6_K_XL.gguf&quot;,
+  &quot;model_path&quot;: &quot;/srv/ai/models/qwen3.6-27b-mtp/Qwen3.6-27B-UD-Q6_K_XL.gguf&quot;,
+  &quot;endpoint&quot;: &quot;http://127.0.0.1:9090/v1/chat/completions&quot;,
+  &quot;proxy&quot;: &quot;http://localhost:10001&quot;,
+  &quot;sampler&quot;: {
+    &quot;temperature&quot;: 0.7,
+    &quot;top_p&quot;: 0.95,
+    &quot;top_k&quot;: 20,
+    &quot;max_tokens&quot;: 32000
+  },
+  &quot;load_command&quot;: &quot;/srv/ai/src/llama.cpp/build/bin/llama-server\n--model /srv/ai/models/qwen3.6-27b-mtp/Qwen3.6-27B-UD-Q6_K_XL.gguf\n--ctx-size 262144\n--cache-type-k f16\n--cache-type-v f16\n--split-mode layer\n--parallel 1\n--spec-type draft-mtp\n--spec-draft-n-max 2\n--host 127.0.0.1 --port 10001 --n-gpu-layers 999 --flash-attn on --cont-batching --jinja --metrics&quot;,
+  &quot;usage&quot;: {
+    &quot;completion_tokens&quot;: 11312,
+    &quot;prompt_tokens&quot;: 838,
+    &quot;total_tokens&quot;: 12150,
+    &quot;prompt_tokens_details&quot;: {
+      &quot;cached_tokens&quot;: 0
+    }
+  },
+  &quot;finish_reason&quot;: &quot;stop&quot;,
+  &quot;performance&quot;: {
+    &quot;wall_secs&quot;: 298.7,
+    &quot;ttft_ms&quot;: 1475.418,
+    &quot;prefill_tps&quot;: 567.9746349847976,
+    &quot;decode_tps&quot;: 42.79383873739593,
+    &quot;prompt_tokens&quot;: 838,
+    &quot;predicted_tokens&quot;: 11312,
+    &quot;prefill_ms&quot;: 1475.418,
+    &quot;decode_ms&quot;: 264337.118
+  },
+  &quot;mtp&quot;: {
+    &quot;draft_n&quot;: 7946,
+    &quot;draft_n_accepted&quot;: 7340,
+    &quot;accept_rate&quot;: 0.924,
+    &quot;enabled&quot;: true
+  },
+  &quot;server_timings&quot;: {
+    &quot;cache_n&quot;: 0,
+    &quot;prompt_n&quot;: 838,
+    &quot;prompt_ms&quot;: 1475.418,
+    &quot;prompt_per_token_ms&quot;: 1.7606420047732696,
+    &quot;prompt_per_second&quot;: 567.9746349847976,
+    &quot;predicted_n&quot;: 11312,
+    &quot;predicted_ms&quot;: 264337.118,
+    &quot;predicted_per_token_ms&quot;: 23.367849893917963,
+    &quot;predicted_per_second&quot;: 42.79383873739593,
+    &quot;draft_n&quot;: 7946,
+    &quot;draft_n_accepted&quot;: 7340
+  },
+  &quot;extraction&quot;: {
+    &quot;extracted&quot;: true,
+    &quot;kind&quot;: &quot;fenced&quot;
+  },
+  &quot;output_file&quot;: &quot;index.html&quot;,
+  &quot;output_bytes&quot;: 32613,
+  &quot;reasoning_bytes&quot;: 621,
+  &quot;wall_secs&quot;: 298.7,
+  &quot;created_utc&quot;: &quot;2026-07-27T04:56:58+00:00&quot;
+}</pre>
+</body></html>
