+++ evals/localmind-landing-page/outputs/chat/meta.json
@@ -1,0 +1,63 @@
+{
+  "test": "localmind-landing-page",
+  "model_slot": "chat",
+  "label": "chat",
+  "model_name": "Qwen3.6-35B-A3B-UD-Q6_K.gguf",
+  "model_path": "/srv/ai/models/qwen3.6-35b-a3b-mtp/Qwen3.6-35B-A3B-UD-Q6_K.gguf",
+  "endpoint": "http://127.0.0.1:9090/v1/chat/completions",
+  "proxy": "http://localhost:10002",
+  "sampler": {
+    "temperature": 0.7,
+    "top_p": 0.95,
+    "top_k": 20,
+    "max_tokens": 32000
+  },
+  "load_command": "/srv/ai/src/llama.cpp/build/bin/llama-server\n--model /srv/ai/models/qwen3.6-35b-a3b-mtp/Qwen3.6-35B-A3B-UD-Q6_K.gguf\n--ctx-size 98304\n--cache-type-k q8_0\n--cache-type-v q8_0\n--parallel 1\n--batch-size 2048 --ubatch-size 2048\n--spec-type draft-mtp --spec-draft-n-max 2\n--host 127.0.0.1 --port 10002 --n-gpu-layers 999 --flash-attn on --cont-batching --jinja --metrics",
+  "usage": {
+    "completion_tokens": 13455,
+    "prompt_tokens": 838,
+    "total_tokens": 14293,
+    "prompt_tokens_details": {
+      "cached_tokens": 0
+    }
+  },
+  "finish_reason": "stop",
+  "performance": {
+    "wall_secs": 101.5,
+    "ttft_ms": 1242.693,
+    "prefill_tps": 674.3419332047416,
+    "decode_tps": 134.51759740113803,
+    "prompt_tokens": 838,
+    "predicted_tokens": 13455,
+    "prefill_ms": 1242.693,
+    "decode_ms": 100024.088
+  },
+  "mtp": {
+    "draft_n": 9294,
+    "draft_n_accepted": 8808,
+    "accept_rate": 0.948,
+    "enabled": true
+  },
+  "server_timings": {
+    "cache_n": 0,
+    "prompt_n": 838,
+    "prompt_ms": 1242.693,
+    "prompt_per_token_ms": 1.4829272076372315,
+    "prompt_per_second": 674.3419332047416,
+    "predicted_n": 13455,
+    "predicted_ms": 100024.088,
+    "predicted_per_token_ms": 7.433971609067261,
+    "predicted_per_second": 134.51759740113803,
+    "draft_n": 9294,
+    "draft_n_accepted": 8808
+  },
+  "extraction": {
+    "extracted": true,
+    "kind": "fenced"
+  },
+  "output_file": "index.html",
+  "output_bytes": 36470,
+  "reasoning_bytes": 882,
+  "wall_secs": 101.5,
+  "created_utc": "2026-07-27T04:51:59+00:00"
+}
