+++ evals/localmind-landing-page/outputs/gemma-31b/meta.json
@@ -1,0 +1,63 @@
+{
+  "test": "localmind-landing-page",
+  "model_slot": "gemma-31b",
+  "label": "gemma-31b",
+  "model_name": "gemma-4-31B-it-qat-UD-Q4_K_XL.gguf",
+  "model_path": "/srv/ai/models/gemma-4-31b/gemma-4-31B-it-qat-UD-Q4_K_XL.gguf",
+  "endpoint": "http://127.0.0.1:9090/v1/chat/completions",
+  "proxy": "http://localhost:10010",
+  "sampler": {
+    "temperature": 0.7,
+    "top_p": 0.95,
+    "top_k": 20,
+    "max_tokens": 32000
+  },
+  "load_command": "/srv/ai/src/llama.cpp/build/bin/llama-server\n--model /srv/ai/models/gemma-4-31b/gemma-4-31B-it-qat-UD-Q4_K_XL.gguf\n--ctx-size 131072\n--cache-type-k q8_0 --cache-type-v q8_0\n--parallel 1\n--batch-size 2048 --ubatch-size 2048\n--model-draft /srv/ai/models/gemma-4-31b/mtp-draft/gemma4-31B-it-assistant-Q8_0.gguf\n--spec-type draft-mtp --spec-draft-n-max 3 --n-gpu-layers-draft 99\n--host 127.0.0.1 --port 10010 --n-gpu-layers 999 --flash-attn on --cont-batching --jinja --metrics",
+  "usage": {
+    "completion_tokens": 8925,
+    "prompt_tokens": 847,
+    "total_tokens": 9772,
+    "prompt_tokens_details": {
+      "cached_tokens": 0
+    }
+  },
+  "finish_reason": "stop",
+  "performance": {
+    "wall_secs": 342.4,
+    "ttft_ms": 1465.379,
+    "prefill_tps": 578.0074642805718,
+    "decode_tps": 28.158692763244982,
+    "prompt_tokens": 847,
+    "predicted_tokens": 8925,
+    "prefill_ms": 1465.379,
+    "decode_ms": 316953.634
+  },
+  "mtp": {
+    "draft_n": 13599,
+    "draft_n_accepted": 4392,
+    "accept_rate": 0.323,
+    "enabled": true
+  },
+  "server_timings": {
+    "cache_n": 0,
+    "prompt_n": 847,
+    "prompt_ms": 1465.379,
+    "prompt_per_token_ms": 1.7300814639905548,
+    "prompt_per_second": 578.0074642805718,
+    "predicted_n": 8925,
+    "predicted_ms": 316953.634,
+    "predicted_per_token_ms": 35.51301221288516,
+    "predicted_per_second": 28.158692763244982,
+    "draft_n": 13599,
+    "draft_n_accepted": 4392
+  },
+  "extraction": {
+    "extracted": true,
+    "kind": "fenced"
+  },
+  "output_file": "index.html",
+  "output_bytes": 29927,
+  "reasoning_bytes": 2056,
+  "wall_secs": 342.4,
+  "created_utc": "2026-07-27T04:50:18+00:00"
+}
