+++ evals/localmind-landing-page/summary.html
@@ -24,7 +24,7 @@ border:1px solid var(--border)}.mtp-on{color:var(--teal);border-color:rgba(0,212
 </style></head>
 <body>
 <h1>localmind-landing-page</h1>
-<p class="sub">Model-eval comparison &middot; 1 run(s) &middot; sorted by objective score, then decode speed</p>
+<p class="sub">Model-eval comparison &middot; 5 run(s) &middot; sorted by objective score, then decode speed</p>
 <div class="wrap"><table>
   <thead><tr>
     <th class="l">Model / weights</th><th>Objective</th><th>MTP</th>
@@ -47,8 +47,68 @@ border:1px solid var(--border)}.mtp-on{color:var(--teal);border-color:rgba(0,212
       <td class="name">2026-07-27</td>
       <td class="l"><a href="outputs/coding/index.html">output</a></td>
     </tr>
+    <tr>
+      <td class="l"><a href="outputs/big/run.html"><span class="model">big</span></a>
+        <div class="name">Qwen3.6-27B-UD-Q6_K_XL.gguf</div></td>
+      <td><span class="score">27/27</span> <span class="name">100%</span></td>
+      <td><span class="pill mtp-on">on</span> <span class="name">92%</span></td>
+      <td>32613 B</td>
+      <td>11312</td>
+      <td>1475.4 ms</td>
+      <td>568.0</td>
+      <td>42.8</td>
+      <td>298.7 s</td>
+      <td>stop</td>
+      <td class="name">2026-07-27</td>
+      <td class="l"><a href="outputs/big/index.html">output</a></td>
+    </tr>
+    <tr>
+      <td class="l"><a href="outputs/gemma-31b/run.html"><span class="model">gemma-31b</span></a>
+        <div class="name">gemma-4-31B-it-qat-UD-Q4_K_XL.gguf</div></td>
+      <td><span class="score">27/27</span> <span class="name">100%</span></td>
+      <td><span class="pill mtp-on">on</span> <span class="name">32%</span></td>
+      <td>29927 B</td>
+      <td>8925</td>
+      <td>1465.4 ms</td>
+      <td>578.0</td>
+      <td>28.2</td>
+      <td>342.4 s</td>
+      <td>stop</td>
+      <td class="name">2026-07-27</td>
+      <td class="l"><a href="outputs/gemma-31b/index.html">output</a></td>
+    </tr>
+    <tr>
+      <td class="l"><a href="outputs/thinkingcap-27b/run.html"><span class="model">thinkingcap</span></a>
+        <div class="name">ThinkingCap-Qwen3.6-27B-Q6_K-MTP.gguf</div></td>
+      <td><span class="score">25/27</span> <span class="name">93%</span></td>
+      <td><span class="pill mtp-on">on</span> <span class="name">93%</span></td>
+      <td>31703 B</td>
+      <td>11266</td>
+      <td>1320.4 ms</td>
+      <td>634.6</td>
+      <td>45.7</td>
+      <td>248.1 s</td>
+      <td>stop</td>
+      <td class="name">2026-07-27</td>
+      <td class="l"><a href="outputs/thinkingcap-27b/index.html">output</a></td>
+    </tr>
+    <tr>
+      <td class="l"><a href="outputs/chat/run.html"><span class="model">chat</span></a>
+        <div class="name">Qwen3.6-35B-A3B-UD-Q6_K.gguf</div></td>
+      <td><span class="score">23/27</span> <span class="name">85%</span></td>
+      <td><span class="pill mtp-on">on</span> <span class="name">95%</span></td>
+      <td>36470 B</td>
+      <td>13455</td>
+      <td>1242.7 ms</td>
+      <td>674.3</td>
+      <td>134.5</td>
+      <td>101.5 s</td>
+      <td>stop</td>
+      <td class="name">2026-07-27</td>
+      <td class="l"><a href="outputs/chat/index.html">output</a></td>
+    </tr>
   </tbody>
 </table></div>
-<p class="foot">Generated 2026-07-27T04:39:23+00:00 by scripts/eval-summary.py.
+<p class="foot">Generated 2026-07-27T05:04:51+00:00 by scripts/eval-summary.py.
 TTFT / throughput are server-side llama.cpp timings (prefill = prompt_ms; decode = predicted_per_second).</p>
 </body></html>
